Add Policy

+++ backend/resources/views/components/comment-form.blade.php
@@ -13,22 +13,22 @@
             キャンセル
         </button>
     </form>
-    <div id="commentList">
-        @foreach ($road->roadComments as $comment)
-            <div class="flex">
-                <div class="font-medium mr-4">{{$comment->user->name}}</div>
-                <div>{{$comment->created_at->format('Y-m-d')}}</div>
-                <form id="delete_comment_{{$comment->id}}" method="post" action="{{ route('user.road.comment.destroy', ['road' => $comment->road_id, 'comment' => $comment->id ]) }}">
-                    @csrf
-                    @method('delete')
-                    <input value="{{ $comment->id }}" type="hidden" name="comment_id" />
-                    <input value="{{ $comment->road->id }}" type="hidden" name="road_id" />
+    @foreach ($road->roadComments as $comment)
+        <div class="flex">
+            <div class="font-medium mr-4">{{$comment->user->name}}</div>
+            <div>{{$comment->created_at->format('Y-m-d')}}</div>
+            <form id="delete_comment_{{$comment->id}}" method="post" action="{{ route('user.road.comment.destroy', ['road' => $comment->road_id, 'comment' => $comment->id ]) }}">
+                @csrf
+                @method('delete')
+                <input value="{{ $comment->id }}" type="hidden" name="comment_id" />
+                <input value="{{ $comment->road->id }}" type="hidden" name="road_id" />
+                @can('delete', $comment)
                     <a href="#" data-id="{{ $comment->id }}" onclick="deleteComment(this)" class="border-0 px-4"><i class="far fa-trash-alt"></i></a>
-                </form>
-            </div>
-            <div class="text-xl">{{$comment->comment}}</div>
-        @endforeach
-    </div>
+                @endcan
+            </form>
+        </div>
+        <div class="text-xl">{{$comment->comment}}</div>
+    @endforeach
 </div>
 
 <script>
